requirements.txt
Flask
Pillow
scikit-image
boto3
wheel
mod_wsgi
